lambda/requirements.txt
boto3==1.9.216
ask-sdk-core==1.11.0
ask-sdk-s3-persistence-adapter==1.0.0
setuptools_scm
pyyaml
pydantic
matplotlib